+++ Cerberus.pro
@@ -13,27 +13,27 @@ CONFIG += c++17
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
-    database.cpp \
-    dialoglogin.cpp \
-    main.cpp \
-    mainwindow.cpp \
-    masterpassword.cpp \
-    password_generator.cpp
+    src/database.cpp \
+    src/dialoglogin.cpp \
+    src/main.cpp \
+    src/mainwindow.cpp \
+    src/masterpassword.cpp \
+    src/password_generator.cpp
 
 
 include ($$PWD/Qt-Secret/src/Qt-Secret.pri)
 
 HEADERS += \
-    database.h \
-    dialoglogin.h \
-    mainwindow.h \
-    masterpassword.h \
-    password_generator.h
+    headers/database.h \
+    headers/dialoglogin.h \
+    headers/mainwindow.h \
+    headers/masterpassword.h \
+    headers/password_generator.h
 
 
 FORMS += \
-    dialoglogin.ui \
-    mainwindow.ui
+    forms/dialoglogin.ui \
+    forms/mainwindow.ui
 
 # Default rules for deployment.
 qnx: target.path = /tmp/$${TARGET}/bin
@@ -41,7 +41,7 @@ else: unix:!android: target.path = /opt/$${TARGET}/bin
 !isEmpty(target.path): INSTALLS += target
 
 RESOURCES += \
-    arrows.qrc \
-    icons.qrc \
-    optionsIcons.qrc \
-    resources.qrc
+    resources/arrows.qrc \
+    resources/icons.qrc \
+    resources/optionsIcons.qrc \
+    resources/resources.qrc
