+++ resources/styles.qss
