+++ src/main.cpp
@@ -1,6 +1,6 @@
 #include <QApplication>
-#include "mainwindow.h"
-#include "dialoglogin.h"
+#include "headers/mainwindow.h"
+#include "headers/dialoglogin.h"
 
 int main(int argc, char *argv[])
 {
@@ -12,7 +12,7 @@ int main(int argc, char *argv[])
         window.resize(1000, 700);
         window.setStyleSheet("QMainWindow { background-color: #121212; }");
         window.setWindowTitle("Cerberus");
-        window.setWindowIcon(QIcon(":/mainIcon.png"));
+        window.setWindowIcon(QIcon("D:/Cerberus/qt/Cerberus/resources/mainIcon.png"));
         window.show();
 
         return app.exec();
